LPS-55492 move to module
