+++ modules/portal/portal-scripting-executor-groovy/test/unit/com/liferay/portal/scripting/groovy/internal/dependencies/binding-input.groovy
@@ -1,3 +1,3 @@
-package com.liferay.portal.scripting.groovy.dependencies
+package com.liferay.portal.scripting.groovy.internal.dependencies
 
 def simple = "Using a $variable passed through the binding mechanism."
