+++ modules/portal/portal-scripting-executor-groovy/test/unit/com/liferay/portal/scripting/groovy/internal/dependencies/runtime-error.groovy
@@ -1,0 +1,4 @@
+package com.liferay.portal.scripting.groovy.internal.dependencies
+
+throw new RuntimeException()
+
