+++ modules/portal/portal-scripting-executor-groovy/test/unit/com/liferay/portal/scripting/groovy/internal/dependencies/simple.groovy
