--- portal-impl/test/unit/com/liferay/portal/scripting/groovy/dependencies/runtime-error.groovy
@@ -1,4 +1,0 @@
-package com.liferay.portal.scripting.groovy.dependencies
-
-throw new RuntimeException()
-
